extracted importProjectAndMoveToMyProjects out of ImportLocalActions.js to ImportLocalHelpers.js
